Ruling out underscores from domain names
Per RFC, underscores are disallowed in domain names. However, sometimes they've appeared in *subdomains* which could form part of the email address. But this has also been under scrutiny from the CA/Browser Forum with them voting to drop support for underscores in DNS names: https://cabforum.org/2018/11/12/ballot-sc-12-sunset-of-underscores-in-dnsnames/

This is becoming important now as we migrate data from Table Storage to SQL Azure and *every* domain on an email address gets an entry in the domain table which has a CHECK constraint on it to limit invalid data. Plus, we're just not seeing this on legitimate addresses. If this turns out to be a false positive in the future, it will be *extraoirindarily* rare and only effect people whose lives are probably already hell when they try to use the address on websites anywway!